feat(themes): :sparkles: break out ThemeTester into its own component
